requirements.txt
Flask==2.3.3
kafka-python==2.0.2
requests==2.31.0
python-dateutil==2.8.2
